Feature gate embedding and rerank code
